huge commit: changing one parameter in the database
sqlite is much, much faster if we don't close the connection at every
single transaction. But to ensure retrocompatibility with kirbybase, we
need to pass the "imgdb" path to both the constructor of the database
interface, and to its methods later on.

Will remove this in the future once we will all have switched to sqlite.